Fix inheritance issue with TaxCore
So when building a layout/view, you need to tell the layout engine
whether its a part of something else or it's just built alone (in
memory). This determines whether or not inheritance rules will kick in.
Since ReceiptExtractor should look the same no matter what, we're
building it in memory and its parent view should be null.

Thank you for attending my Ted talk. May you have a wonderful evening.